bug fixes and updates before 1.4.8 release

+++ upload/system/library/cache.php
@@ -10,7 +10,9 @@ final class Cache {
 				$time = substr(strrchr($file, '.'), 1);
 
       			if ($time < time()) {
-					unlink($file);
+					if (file_exists(DIR_CACHE . $file)) {
+						unlink($file);
+					}
       			}
     		}
 		}
